CAN: Move some FDCAN definitions into a module to share with BXCAN.
